backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
langchain
openai
